update other platforms, and enhance gui

+++ minibae/src/BAE_Source/Platform/BAE_API_Ansi.c
@@ -417,7 +417,7 @@ int32_t BAE_FileDelete(void *fileName)
 
 // Open a file
 // Return -1 if error, otherwise file handle
-int32_t BAE_FileOpenForRead(void *fileName)
+intptr_t BAE_FileOpenForRead(void *fileName)
 {
 
 	if (fileName)
@@ -426,7 +426,7 @@ int32_t BAE_FileOpenForRead(void *fileName)
 	   return _open((char *)fileName, _O_RDONLY | _O_BINARY);
 #elif USE_ANSI_IO
        FILE *fp = fopen((char *)fileName, "rb");
-       return (int32_t)fp;
+       return (intptr_t)fp;
 #elif USE_WINDOWS_IO
 		HANDLE	file;
 
@@ -442,13 +442,13 @@ int32_t BAE_FileOpenForRead(void *fileName)
 			DWORD	lastErr = GetLastError();
 			return -1;
 		}
-		return (int32_t)file;
+		return (intptr_t)file;
 #endif
 	}
 	return -1;
 }
 
-int32_t BAE_FileOpenForWrite(void *fileName)
+intptr_t BAE_FileOpenForWrite(void *fileName)
 {
 
 	if (fileName)
@@ -457,7 +457,7 @@ int32_t BAE_FileOpenForWrite(void *fileName)
 		return _open((char *)fileName, _O_WRONLY | _O_CREAT | _O_TRUNC | _O_BINARY);
 #elif USE_ANSI_IO
 		FILE *fp = fopen((char *)fileName, "wb");
-		return (int32_t)fp;;
+		return (intptr_t)fp;;
 #elif USE_WINDOWS_IO
 		HANDLE	file;
 
@@ -469,14 +469,14 @@ int32_t BAE_FileOpenForWrite(void *fileName)
 			DWORD	lastErr = GetLastError();
 			return -1;
 		}
-		return (int32_t)file;
+		return (intptr_t)file;
 
 #endif
 	}
 	return -1;
 }
 
-int32_t BAE_FileOpenForReadWrite(void *fileName)
+intptr_t BAE_FileOpenForReadWrite(void *fileName)
 {
 	if (fileName)
 	{
@@ -484,7 +484,7 @@ int32_t BAE_FileOpenForReadWrite(void *fileName)
 		return _open((char *)fileName, _O_RDWR | _O_BINARY);
 #elif USE_ANSI_IO
 		FILE *fp = fopen((char *)fileName, "r+b" /*"arb"*/ /*"wrb"*/);
-		return (int32_t)fp;
+		return (intptr_t)fp;
 #elif USE_WINDOWS_IO
 		HANDLE	file;
 
@@ -499,7 +499,7 @@ int32_t BAE_FileOpenForReadWrite(void *fileName)
 			DWORD	lastErr = GetLastError();
 			return -1;
 		}
-		return (int32_t)file;
+		return (intptr_t)file;
 
 #endif
 	}
@@ -507,7 +507,7 @@ int32_t BAE_FileOpenForReadWrite(void *fileName)
 }
 
 // Close a file
-void BAE_FileClose(int32_t fileReference)
+void BAE_FileClose(intptr_t fileReference)
 {
 #if  USE_UNIX_IO
 	_close(fileReference);
@@ -520,7 +520,7 @@ void BAE_FileClose(int32_t fileReference)
 
 // Read a block of memory from a file.
 // Return -1 if error, otherwise length of data read.
-int32_t BAE_ReadFile(int32_t fileReference, void *pBuffer, int32_t bufferLength)
+int32_t BAE_ReadFile(intptr_t fileReference, void *pBuffer, int32_t bufferLength)
 {
 	if (pBuffer && bufferLength)
 	{
@@ -544,7 +544,7 @@ int32_t BAE_ReadFile(int32_t fileReference, void *pBuffer, int32_t bufferLength)
 
 // Write a block of memory from a file
 // Return -1 if error, otherwise length of data written.
-int32_t BAE_WriteFile(int32_t fileReference, void *pBuffer, int32_t bufferLength)
+int32_t BAE_WriteFile(intptr_t fileReference, void *pBuffer, int32_t bufferLength)
 {
 	if (pBuffer && bufferLength)
 	{
@@ -571,7 +571,7 @@ int32_t BAE_WriteFile(int32_t fileReference, void *pBuffer, int32_t bufferLength
 
 // set file position in absolute file byte position
 // Return -1 if error, otherwise 0.
-int32_t BAE_SetFilePosition(int32_t fileReference, uint32_t filePosition)
+int32_t BAE_SetFilePosition(intptr_t fileReference, uint32_t filePosition)
 {
 #if  USE_UNIX_IO
 	return (_lseek(fileReference, filePosition, SEEK_SET) == -1) ? -1 : 0;
@@ -587,7 +587,7 @@ int32_t BAE_SetFilePosition(int32_t fileReference, uint32_t filePosition)
 }
 
 // get file position in absolute file bytes
-uint32_t BAE_GetFilePosition(int32_t fileReference)
+uint32_t BAE_GetFilePosition(intptr_t fileReference)
 {
 #if USE_UNIX_IO
 	return _lseek(fileReference, 0, SEEK_CUR);
@@ -599,7 +599,7 @@ uint32_t BAE_GetFilePosition(int32_t fileReference)
 }
 
 // get length of file
-uint32_t BAE_GetFileLength(int32_t fileReference)
+uint32_t BAE_GetFileLength(intptr_t fileReference)
 {
 	uint32_t pos = 0;
 	int val = 0;
@@ -622,7 +622,7 @@ uint32_t BAE_GetFileLength(int32_t fileReference)
 }
 
 // set the length of a file. Return 0, if ok, or -1 for error
-int BAE_SetFileLength(int32_t fileReference, uint32_t newSize)
+int BAE_SetFileLength(intptr_t fileReference, uint32_t newSize)
 {
 #if USE_UNIX_IO
 	return _chsize(fileReference, newSize);
